# Use RMM's DeviceBuffer (in cudf-merge and elsewhere)