--- a/benchmarks/cudf-merge.py
+++ b/benchmarks/cudf-merge.py
@@ -16,6 +16,8 @@ from dask.utils import format_bytes, format_time, parse_bytes
 
 import cudf
 import cupy
+import numba
+import numba.cuda
 import numpy as np
 import rmm
 import ucp
@@ -33,6 +35,15 @@ async def send_df(ep, df):
         await ep.send(frame)
 
 
+def device_array(shape, dtype):
+    dtype = np.dtype(dtype)
+    nbytes = np.prod(shape) * dtype.itemsize
+    buf = rmm.DeviceBuffer(size=nbytes)
+    arr = cupy.asarray(buf).view(dtype).reshape(shape)
+    arr = numba.cuda.as_cuda_array(arr)
+    return arr
+
+
 async def recv_df(ep):
     header_nbytes = np.empty((1,), dtype=np.uint64)
     await ep.recv(header_nbytes)
@@ -41,7 +52,7 @@ async def recv_df(ep):
     header = pickle.loads(header)
 
     frames = [
-        rmm.device_array(iface["shape"], dtype=np.dtype(iface["typestr"]))
+        device_array(iface["shape"], dtype=iface["typestr"])
         for iface in header["frame_ifaces"]
     ]
     for frame in frames: