Use RMM `DeviceBuffer` for allocations.