@jakirkham 2020-01-15
For both of the following, ran this command on latest nightlies:

```
UCXPY_IFNAME="enp1s0f0" UCX_TLS=tcp,cuda_copy,sockcm python ucx-py/benchmarks/cudf-merge.py --chunks-per-dev 4 --chunk-size 10000 --rmm-init-pool-size 100
```

Without this change I see this performance:

```
cudf merge benchmark
----------------------------
device(s)      | [0]
chunks-per-dev | 4
rows-per-chunk | 10000
data-processed | 1.28 MB
frac-match     | 0.3
============================
Wall-clock     | 48.53 ms
Bandwidth      | 11.62 MB/s
Throughput     | 26.38 MB/s
```

With this change I see this performance:

```
cudf merge benchmark
----------------------------
device(s)      | [0]
chunks-per-dev | 4
rows-per-chunk | 10000
data-processed | 1.28 MB
frac-match     | 0.3
============================
Wall-clock     | 42.80 ms
Bandwidth      | 16.21 MB/s
Throughput     | 29.91 MB/s
```