@pentschev 2020-01-15
@jakirkham could you also share results for larger problems (perhaps increasing `chunks-per-dev` or `rows-per-chunk`, or maybe both)?

I see a gain of performance here, but I'm not sure I understand why that happens and how you came to the conclusion that replacing RMM for CuPy would have that effect. Could you elaborate on that?