@jakirkham 2020-01-15
Just for clarity both cases are [allocating through RMM]( https://github.com/rapidsai/ucx-py/blob/593fc8d7d1e4b9997be0816466b7ed9a27c53cef/benchmarks/cudf-merge.py#L172-L173 ). It's just a difference in APIs. Am making this change w.r.t. the discussion in issue ( https://github.com/rapidsai/rmm/issues/180 ) to remove this API from RMM.

That said, I was curious of the performance impact. So ran things to see. Agree it is not big (nor would I expect it to be). I don't expect there to be a particularly noticeable scale difference as we are using the same underlying allocator.